<1.2 upgrades> Pull in changes for stg-hotfix 2.0.28.4 and 2.0.28.5 plus additional fix for Bug 973740

+++ upgrades/lib/ose-upgrade/node/upgrades/1/gears/28-v2carts/migrate-v2.rb
@@ -30,6 +30,7 @@ require 'openshift-origin-node/utils/sdk'
 require 'openshift-origin-node/utils/cgroups'
 require 'openshift-origin-node/utils/application_state'
 require 'openshift-origin-node/utils/environ'
+require 'openshift-origin-node/utils/selinux'
 require 'openshift-origin-common'
 require 'net/http'
 require 'uri'
@@ -62,7 +63,7 @@ module OpenShift
         cart_status = do_control('status', cartridge)
 
         cart_status_msg = "[OK]"
-        if cart_status !~ /running|enabled|Tail of JBoss/i
+        if cart_status !~ /running|enabled|Tail of JBoss|status output from the mock cartridge/i
           problem = true
           cart_status_msg = "[PROBLEM]"
         end
@@ -131,6 +132,8 @@ module OpenShiftMigration
         begin
           progress.log 'Beginning V1 -> V2 migration'
 
+          progress.init_store
+
           if detect_malformed_gear(progress, gear_home)
             progress.log 'Deleting migration metadata because this gear appears not to have any cartridges'
             progress.done
@@ -159,7 +162,10 @@ module OpenShiftMigration
           start_gear(progress, params[:hostname], params[:uuid])
 
           validate_gear(progress, params[:uuid], params[:gear_home])
-          cleanup(progress, params[:gear_home])
+
+          if progress.complete? 'validate_gear'
+            cleanup(progress, params[:gear_home])
+          end
 
         rescue OpenShift::Utils::ShellExecutionException => e
           progress.log %Q(#{e.message} stdout => \n #{e.stdout} stderr => \n #{e.stderr})
@@ -202,9 +208,22 @@ module OpenShiftMigration
 
         if progress.incomplete? 'inspect_gear_state'
           app_state = File.join(gear_home, 'app-root', 'runtime', '.state')
-          save_state = File.join(gear_home, 'app-root', 'runtime', PREMIGRATION_STATE)
+          runtime_dir = File.join(gear_home, 'app-root', 'runtime')
+          save_state = File.join(runtime_dir, PREMIGRATION_STATE)
 
-          FileUtils.cp(app_state, save_state)
+          if !File.exists?(runtime_dir)
+            FileUtils.mkpath(runtime_dir)
+          end
+
+          if File.exists? app_state
+            FileUtils.cp(app_state, save_state)
+          else
+            IO.write(save_state, 'stopped')
+            mcs_label = OpenShift::Utils::SELinux.get_mcs_label(uuid)
+            user = OpenShift::UnixUser.from_uuid(uuid)
+            PathUtils.oo_chown(user.uid, user.gid, save_state)
+            OpenShift::Utils::SELinux.set_mcs_label(mcs_label, save_state)
+          end
 
           premigration_state = OpenShift::Utils::MigrationApplicationState.new(uuid, PREMIGRATION_STATE)
           progress.log "Pre-migration state: #{premigration_state.value}"
@@ -339,7 +358,12 @@ module OpenShiftMigration
           Dir.glob(File.join(gear_home, '.env', '*')).each do |entry|
             next if File.basename(entry) == 'TYPELESS_TRANSLATED_VARS'
 
-            content = IO.read(entry).chomp
+            begin
+              content = IO.read(entry).chomp
+            rescue Exception => e
+              progress.log "Error reading from #{entry}; skipping."
+              next
+            end
 
             if content =~ /^export /
               index          = content.index('=')
@@ -426,6 +450,12 @@ module OpenShiftMigration
         end
 
         if repo.exists? && progress.incomplete?("reconfigure_git_repo")
+          hooks_dir = File.join(repo.path, 'hooks')
+
+          if !File.exists?(hooks_dir)
+            FileUtils.mkpath(hooks_dir)
+          end
+
           repo.configure
           progress.mark_complete('reconfigure_git_repo')
         end
@@ -607,8 +637,7 @@ module OpenShiftMigration
             problem, status = cart_model.gear_status
 
             if problem
-              progress.log "Leaving migration metadata in place due to problem detected with gear status:\n#{status}"
-              return
+              progress.log "Problem detected with gear status:\n#{status}"
             end
           end
 
@@ -617,11 +646,9 @@ module OpenShiftMigration
       end
 
       def self.cleanup(progress, gear_home)
-        if progress.complete? 'validate_gear'
-          progress.log 'Cleaning up after migration'
-          FileUtils.rm_f(File.join(gear_home, 'app-root', 'runtime', PREMIGRATION_STATE))
-          progress.done
-        end
+        progress.log 'Cleaning up after migration'
+        FileUtils.rm_f(File.join(gear_home, 'app-root', 'runtime', PREMIGRATION_STATE))
+        progress.done
       end
 
       ## These will replace their un-versioned counterparts in version 2.0.29
@@ -649,16 +676,26 @@ module OpenShiftMigration
 
         begin
           progress.log "Beginning #{version} migration for #{uuid}"
-          inspect_gear_state(progress, uuid, gear_home)
 
+          inspect_gear_state(progress, uuid, gear_home)
           migrate_cartridges_v2v2(progress, gear_home, gear_env, uuid, hostname)
 
-          validate_gear(progress, uuid, gear_home)
-          cleanup(progress, gear_home)
+          if progress.has_instruction?('validate_gear')
+            validate_gear(progress, uuid, gear_home)
+
+            if progress.complete? 'validate_gear'
+              cleanup(progress, gear_home)
+            end
+          else
+            cleanup(progress, gear_home)
+          end
 
           total_time = (Time.now.to_f * 1000).to_i - start_time
           progress.log "***time_migrate_on_node_measured_from_node=#{total_time}***"
-        rescue => e
+        rescue OpenShift::Utils::ShellExecutionException => e
+          progress.log %Q(#{e.message} stdout => \n #{e.stdout} stderr => \n #{e.stderr})
+          exitcode = 1
+        rescue Exception => e
           progress.log "Caught an exception during internal migration steps: #{e.message}"
           progress.log e.backtrace.join("\n")
           exitcode = 1
@@ -706,28 +743,36 @@ module OpenShiftMigration
                 next
               end
 
-              if next_manifest.compatible_versions.include?(cartridge_version)
-                progress.log "Compatible migration of cartridge #{ident}"
-                compatible_migration(progress, cartridge_model, next_manifest, version, cartridge_path, user)
-              else
-                stop_gear(progress, hostname, uuid) unless restart_required
-                restart_required = true
+              if progress.incomplete? "#{name}_migrate"
+                progress.set_instruction('validate_gear')
 
-                progress.log "Incompatible migration of cartridge #{ident}"
-                incompatible_migration(progress, cartridge_model, next_manifest, version, cartridge_path, user)
+                if next_manifest.compatible_versions.include?(cartridge_version)
+                  progress.log "Compatible migration of cartridge #{ident}"
+                  compatible_migration(progress, cartridge_model, next_manifest, cartridge_path, user)
+                else
+                  stop_gear(progress, hostname, uuid) unless progress.has_instruction?('restart_gear')
+                  progress.set_instruction('restart_gear')
+
+                  progress.log "Incompatible migration of cartridge #{ident}"
+                  incompatible_migration(progress, cartridge_model, next_manifest, version, cartridge_path, user)
+                end
+
+                progress.mark_complete("#{name}_migrate")
               end
 
-              next_ident = OpenShift::Runtime::Manifest.build_ident(manifest.cartridge_vendor,
-                                                                    manifest.name,
-                                                                    manifest.version,
-                                                                    next_manifest.cartridge_version)
-              IO.write(ident_path, next_ident, 0, mode: 'w', perms: 0666)
-              progress.mark_complete("#{manifest.name}_update_ident")
+              if progress.incomplete? "#{name}_rebuild_ident"
+                next_ident = OpenShift::Runtime::Manifest.build_ident(manifest.cartridge_vendor,
+                                                                      manifest.name,
+                                                                      manifest.version,
+                                                                      next_manifest.cartridge_version)
+                IO.write(ident_path, next_ident, 0, mode: 'w', perms: 0666)
+                progress.mark_complete("#{name}_rebuild_ident")
+              end
             end
           end
         end
 
-        if restart_required
+        if progress.has_instruction?('restart_gear')
           restart_start_time = (Time.now.to_f * 1000).to_i
           start_gear(progress, hostname, uuid)
           restart_time = (Time.now.to_f * 1000).to_i - restart_start_time
@@ -736,12 +781,12 @@ module OpenShiftMigration
       end
 
       # Simple change that does not require the gear to be restarted
-      def self.compatible_migration(progress, cart_model, next_manifest, version, target, user)
+      def self.compatible_migration(progress, cart_model, next_manifest, target, user)
         OpenShift::CartridgeRepository.overlay_cartridge(next_manifest, target)
 
         # No ERB's are rendered for fast migrations
         FileUtils.rm_f cart_model.processed_templates(next_manifest)
-        progress.mark_complete("#{name}_remove_erb")
+        progress.mark_complete("#{next_manifest.name}_remove_erb")
 
         cart_model.unlock_gear(next_manifest) do |m|
           cart_model.secure_cartridge(next_manifest.short_name, user.uid, user.gid, target)
@@ -749,8 +794,12 @@ module OpenShiftMigration
       end
 
       def self.incompatible_migration(progress, cart_model, next_manifest, version, target, user)
+        FileUtils.rm_f cart_model.setup_rewritten(next_manifest)
+
         OpenShift::CartridgeRepository.overlay_cartridge(next_manifest, target)
         cart_model.secure_cartridge(next_manifest.short_name, user.uid, user.gid, target)
+
+        name = next_manifest.name
 
         cart_model.unlock_gear(next_manifest) do |m|
           if progress.incomplete? "#{name}_setup"
